Rename upsertPersonsStep => processPersonsStep
