+++ plugin-server/src/worker/ingestion/event-pipeline/1-emitToBufferStep.ts
@@ -20,7 +20,7 @@ export async function emitToBufferStep(
         await runner.hub.eventsProcessor.produceEventToBuffer(event)
         return null
     } else {
-        return runner.nextStep('upsertPersonsStep', event, person)
+        return runner.nextStep('processPersonsStep', event, person)
     }
 }
