+++ plugin-server/src/worker/ingestion/event-pipeline/2-processPersonsStep.ts
@@ -14,7 +14,7 @@ export interface ForwardedPersonData {
     }
 }
 
-export async function upsertPersonsStep(
+export async function processPersonsStep(
     runner: EventPipelineRunner,
     event: PluginEvent,
     person: Person | undefined
